parse: keep track of authorized_keys options
it also keeps track of known_hosts hostnames and markers in the same
field. even though these aren't a part of the spec, they are a superset
of the spec, so my current line of thought is to handle them here
instead of having special logic in read_keys.

fixes #7